差个反序列化加载的有问题

+++ Unity/Assets/Scripts/Editor/Tree/EditorNode/Task/WaitEditorNode.cs
@@ -6,7 +6,7 @@ namespace ET
     [NodeMenuItem("Task/Wait", typeof(TreeGraph))]
     public class WaitEditorNode : TaskEditorNode
     {
-        public float Seconds;
+        public long Seconds;
         
         public override object Init(Dictionary<string, object> blackboard)
         {
